feature(twig): reworked build script to use rollup and output umd
