+++ packages/twig-new/package.json
@@ -7,13 +7,16 @@
     "dist/**/*"
   ],
   "scripts": {
-    "build:lib": "vite build",
+    "build:lib": "rollup -c",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^1.5.0",
     "@modyfi/vite-plugin-yaml": "^1.1.0",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-swc": "^0.3.1",
     "@storybook/addon-essentials": "^8.1.9",
     "@storybook/addon-interactions": "^8.1.9",
     "@storybook/addon-links": "^8.1.9",
@@ -22,6 +25,7 @@
     "@storybook/html": "^8.1.9",
     "@storybook/html-vite": "^8.1.9",
     "@storybook/test": "^8.1.9",
+    "@swc/core": "^1.7.1",
     "drupal-attribute": "^1.0.2",
     "drupal-twig-extensions": "^1.0.0-beta.5",
     "glob": "^10.3.10",
@@ -29,7 +33,6 @@
     "storybook": "^8.1.9",
     "twig": "^1.16.0",
     "vite": "^5.2.0",
-    "vite-plugin-twig-drupal": "^1.3.0",
     "zod": "^3.23.8"
   },
   "dependencies": {
